+++ loinc2hpogui/src/main/java/org/monarchinitiative/loinc2hpo/github/GitHubPoster.java
@@ -8,6 +8,8 @@ import java.net.HttpURLConnection;
 import java.net.URL;
 import java.net.URLConnection;
 import java.nio.charset.StandardCharsets;
+import java.util.List;
+import java.util.stream.Collectors;
 
 
 /**
@@ -42,6 +44,8 @@ public class GitHubPoster {
 
     private String githubLabel = null;
 
+    private List<String> githubLabels = null;
+
     private String githubTitle = null;
 
     private String githubBody = null;
@@ -68,6 +72,11 @@ public class GitHubPoster {
         reformatPayloadWithLabel(githubLabel);
     }
 
+    public void setLabel(List<String> labels) {
+        this.githubLabels = labels.stream().map(JSONValue::escape).collect(Collectors.toList());
+        reformatPayloadWithLabel(this.githubLabels);
+    }
+
 
     /**
      * TODO create our won escape formated (new line, quotation mark etc.
@@ -77,14 +86,56 @@ public class GitHubPoster {
     }
 
 
+    /**
+     * Change a list of string into a Json array
+     * @param labels
+     * @return
+     */
+    private String labelsArray4Json(List<String> labels) {
+
+        if (labels.size() == 1) {
+            return "\"" + labels.get(0) + "\"";
+        }
+
+        if (labels.size() > 1) {
+            StringBuilder sb = new StringBuilder();
+            sb.append("\"" + labels.get(0) + "\"");
+            for (int i = 1; i < labels.size(); i++) {
+                sb.append(", \"" + labels.get(i) + "\"");
+            }
+            return sb.toString();
+        }
+
+        return null;
+
+    }
+
+    public String debugLabelsArray4Json() {
+        return labelsArray4Json(this.githubLabels);
+    }
+
+    public String debugReformatpayloadWithLabel() {
+        return this.payload;
+    }
+
     private void reformatPayloadWithLabel(String label) {
         this.payload = String.format("{\n" +
                         "\"title\": \"%s\",\n" +
                         "\"body\": \"%s\",\n" +
-                        "\"labels\": [ \"%s\"] }",
+                        "\"labels\": [ \"%s\" ] }",
                 JSONValue.escape(this.githubTitle),
                 JSONValue.escape(this.githubBody),
                 JSONValue.escape(this.githubLabel));
+    }
+
+    private void reformatPayloadWithLabel(List<String> labels) {
+        this.payload = String.format("{\n" +
+                        "\"title\": \"%s\",\n" +
+                        "\"body\": \"%s\",\n" +
+                        "\"labels\": [ %s] }",
+                JSONValue.escape(this.githubTitle),
+                JSONValue.escape(this.githubBody),
+                labelsArray4Json(this.githubLabels));
     }